+++ src/Autocomplete/tests/Fixtures/Kernel.php
@@ -82,7 +82,7 @@ final class Kernel extends BaseKernel
             {
                 $container->removeDefinition('doctrine.orm.listeners.pdo_session_handler_schema_listener');
 
-                if (\PHP_VERSION_ID >= 80400 && $container->hasParameter('doctrine.orm.proxy_dir')) {
+                if ($container->hasParameter('doctrine.orm.proxy_dir')) {
                     // Workaround for `RuntimeException: Unable to create the Doctrine Proxy directory "". in vendor/symfony/doctrine-bridge/CacheWarmer/ProxyCacheWarmer.php:49`
                     // when running PHP 8.4 and Doctrine ORM 3.5+.
                     $container->getDefinition('doctrine.orm.default_configuration')
@@ -142,7 +142,7 @@ final class Kernel extends BaseKernel
                 }
             }
 
-            if (\PHP_VERSION_ID >= 80400 && version_compare($doctrineBundleVersion, '2.15.0', '>=') && version_compare($doctrineBundleVersion, '4.0.0', '<')) {
+            if (version_compare($doctrineBundleVersion, '2.15.0', '>=') && version_compare($doctrineBundleVersion, '4.0.0', '<')) {
                 $doctrineConfig['orm']['enable_native_lazy_objects'] = true;
             }
         }