+++ src/LiveComponent/src/DependencyInjection/LiveComponentExtension.php
@@ -203,7 +203,7 @@ final class LiveComponentExtension extends Extension implements PrependExtension
             ->setArguments([
                 new Reference('ux.twig_component.component_factory'),
                 new Reference('property_info'),
-                new Reference('type_info.resolver', ContainerInterface::NULL_ON_INVALID_REFERENCE),
+                new Reference('type_info.resolver'),
             ])
             ->addTag('kernel.reset', ['method' => 'reset'])
         ;