+++ src/LiveComponent/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 3.0.0
 
 - Minimum required Symfony version is now 6.4
-- Minimum required PHP version is now 8.2
+- Minimum required PHP version is now 8.4
 - Remove `csrf` argument from `AsLiveComponent` in favor of same-origin/CORS
 
 ## 2.35